requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
chardet==5.2.0
click==8.1.8
colorama==0.4.6
coverage==7.8.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
pydantic==2.11.1
pydantic_core==2.33.0
pytest==8.3.5
pytest-cov==6.1.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
reportlab==4.3.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uvicorn==0.34.0
